+++ SBA_Serv/Game/FindTheMiddle.py
@@ -1,7 +1,7 @@
 """
 Space Battle Arena is a Programming Game.
 
-Copyright (C) 2012-2015 Michael A. Hawker and Brett Wortzman
+Copyright (C) 2012-2016 Michael A. Hawker and Brett Wortzman
 
 This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version.
 
@@ -41,6 +41,8 @@ class FindTheMiddleGame(BasicGame):
 
     def world_create(self):
         self.midpoint = (int(self.world.width / 2), int(self.world.height / 2))
+
+        super(FindTheMiddleGame, self).world_create()
 
     def player_added(self, player, reason):
         player.time = 0