+++ SBA_Serv/World/WorldMap.py
@@ -1,7 +1,7 @@
 """
 Space Battle Arena is a Programming Game.
 
-Copyright (C) 2012-2015 Michael A. Hawker and Brett Wortzman
+Copyright (C) 2012-2016 Michael A. Hawker and Brett Wortzman
 
 This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version.
 
@@ -220,7 +220,7 @@ class GameWorld(object):
                         obj.body.position[1] -= self.height
 
                     # Apply Gravity for Planets
-                    if obj.explodable:
+                    if obj.gravitable:
                         for influencer in self.__influential:
                             for point in wrappos(obj.body.position, influencer.influence_range, self.size):
                                 if in_circle(influencer.body.position, influencer.influence_range, point):
@@ -277,8 +277,11 @@ class GameWorld(object):
             self.gameerror = True
         logging.debug("Gameloop Ended")
 
-    def getObjectsInArea(self, pos, radius, force=False):
-        logging.debug("Get Objects In Area %s %d (%s) [%d]", repr(pos), radius, repr(force), thread.get_ident())
+    def getObjectsInArea(self, pos, radius, radar=False):
+        """
+        Returns objects within the given radius from the position (even wrapping around edge of world), pass radar = True if for environment
+        """
+        logging.debug("Get Objects In Area %s %d [%d]", repr(pos), radius, thread.get_ident())
         objList = []
         for obj in self.__objects.values():
             for point in wrappos(obj.body.position, radius, self.size):
@@ -307,7 +310,7 @@ class GameWorld(object):
         objData["POSITION"] = intpos(obj.body.position)
         objData["SPEED"] = obj.body.velocity.length
         # TODO: deal with -0.0 case OR match physics library coordinates?
-        objData["DIRECTION"] = -obj.body.velocity.angle_degrees # 30 = -120?, -80 = -10
+        objData["DIRECTION"] = -obj.body.velocity.angle_degrees % 360 # 30 = -120?, -80 = -10
         #objData["VELOCITYDIRECTION"] = obj.velocity.direction
         objData["MAXSPEED"] = obj.body.velocity_limit
         objData["CURHEALTH"] = obj.health.value
@@ -330,7 +333,7 @@ class GameWorld(object):
         #TODO: abstract radar to game level?
         radardata = None
         if radarlevel > 0:
-            objs = self.getObjectsInArea(ship.body.position, ship.radarRange)
+            objs = self.getObjectsInArea(ship.body.position, ship.radarRange, True)
             #TODO: Need to wait lock the removing of ships with accessing...???
             if ship in objs: objs.remove(ship) # Get rid of self...