+++ changelog.md
@@ -1,5 +1,20 @@
 v1.2 : Planned - May 2016 [Season 5]
 ----
+* Added Worm Holes (teleports objects across space)
+* Added Deployable Space Mines
+* Updated Client Angle/Point Methods
+    * Orientations are always integers now and should be between 0-360
+        * Rotate and Steer Command parameters updated
+    * Movement direction can still be a double
+    * Added Test Cases for Point Class
+       * Fixed **getAngleTo** method to use built-in atan2
+    * Added **isCloseTo** box test method
+    * Added **getPointAt** for point projections based on angles and distance
+    * Added **isInEllipse** check method
+    * Added **getClosestMappedPoint** for world bounds assistance
+* Added option for Torpedoes to be effected by gravity
+* Split 'explodable' from 'gravitable' for Entities, two separate object flags now.
+* Separated option for 'showip' in Application settings to decouple from showing statistics, no longer always show IP in Debug mode.
 
 v1.1.0.1111 : 04/21/2016 [Season 4 Release] - Discovery Quest
 ----